practice6

+++ src/main/java/kz/medet/userservice/service/impl/CustomerService.java
@@ -7,6 +7,7 @@ import kz.medet.userservice.entity.CustomerDocument;
 import kz.medet.userservice.exceptions.CustomException;
 import kz.medet.userservice.kafka.KafkaConsumer;
 import kz.medet.userservice.kafka.KafkaProducer;
+import kz.medet.userservice.mapper.CustomerMapper;
 import kz.medet.userservice.repository.CustomerRepository;
 import kz.medet.userservice.repository.CustomerSearchRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,11 +15,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 @Service
 @Transactional
@@ -26,32 +28,122 @@ public class CustomerService {
 
     private final CustomerRepository customerRepository;
     private final KafkaProducer kafkaProducer;
-
     private final KafkaConsumer kafkaConsumer;
     private final BlockingQueue blockingQueue;
-
     private final OrderServiceClient orderServiceClient;
-
     private final CustomerSearchRepository customerSearchRepository;
+    private final CustomerMapper customerMapper;
 
     public CustomerService(CustomerRepository customerRepository,
                            KafkaProducer kafkaProducer,
                            KafkaConsumer kafkaConsumer,
                            BlockingQueue blockingQueue,
                            CustomerSearchRepository customerSearchRepository,
-                           OrderServiceClient orderServiceClient) {
+                           OrderServiceClient orderServiceClient,
+                           CustomerMapper customerMapper) {
         this.customerRepository = customerRepository;
         this.kafkaProducer = kafkaProducer;
         this.kafkaConsumer = kafkaConsumer;
         this.blockingQueue = blockingQueue;
         this.customerSearchRepository = customerSearchRepository;
         this.orderServiceClient = orderServiceClient;
+        this.customerMapper = customerMapper;
     }
+
+    @Transactional(readOnly = true, propagation = Propagation.REQUIRED)
+    public Map<String, Long> compareStreamPerformance() {
+        List<Customer> customers = customerRepository.findAll();
+
+        long startSequential = System.currentTimeMillis();
+        customers.stream()
+                .map(customerMapper::toResponseDto)
+                .toList();
+        long sequentialTime = System.currentTimeMillis() - startSequential;
+
+        long startParallel = System.currentTimeMillis();
+        customers.parallelStream()
+                .map(customerMapper::toResponseDto)
+                .toList();
+        long parallelTime = System.currentTimeMillis() - startParallel;
+
+        System.out.println("Время выполнения (последовательно): " + sequentialTime + " мс");
+        System.out.println("Время выполнения (параллельно): " + parallelTime + " мс");
+
+        Map<String, Long> result = new HashMap<>();
+        result.put("sequential", sequentialTime);
+        result.put("parallel", parallelTime);
+        return result;
+    }
+
+
+    @Transactional(readOnly = true)
+    public List<CustomerDocument> filterCustomers(String query, Long orderId) {
+        return StreamSupport.stream(customerSearchRepository.findAll().spliterator(), false)
+                .filter(c -> (query == null || c.getFirstName().toLowerCase().contains(query.toLowerCase())
+                        || c.getLastName().toLowerCase().contains(query.toLowerCase())))
+                .filter(c -> (orderId == null || c.getOrderId() != null && c.getOrderId().equals(orderId)))
+                .collect(Collectors.toList());
+    }
+
+    @Transactional(readOnly = true, propagation = Propagation.REQUIRED)
+    public Map<String, List<Customer>> groupCustomersByLastName() {
+        List<Customer> customers = customerRepository.findAll();
+
+        return customers.stream()
+                .collect(Collectors.groupingBy(Customer::getLastName));
+    }
+
+
+
 
     @Transactional(readOnly = true, propagation = Propagation.REQUIRED)
     public List<Customer> getAllCustomers() {
         return customerRepository.findAll();
     }
+
+    @Transactional(readOnly = true, propagation = Propagation.REQUIRED)
+    public long countAllCustomers() {
+        return customerRepository.findAll()
+                .stream()
+                .map(customer -> 1L)
+                .reduce(0L, Long::sum);
+    }
+
+    public long countCustomersWithOrders() {
+        return getAllCustomers()
+                .stream()
+                .map(Customer::getOrderId)
+                .filter(Objects::nonNull)
+                .reduce(0L, (sum, orderId) -> sum + 1);
+    }
+
+
+
+
+    public List<CustomerResponseDto> getAllCustomer(String nameQuery, Long orderId, int page, int size) {
+        if (nameQuery == null || nameQuery.isBlank()) {
+            return getAllCustomers()
+                    .parallelStream()
+                    .filter(customer -> orderId == null || customer.getOrderId().equals(orderId))
+                    .map(customerMapper::toResponseDto)
+                    .toList();
+        }
+
+        List<CustomerDocument> customerDocuments = customerSearchRepository
+                .findAllByFirstNameContainingIgnoreCaseOrLastNameContainingIgnoreCase(nameQuery, nameQuery);
+
+        Set<Long> idsSet = customerDocuments.stream()
+                .map(CustomerDocument::getId)
+                .map(Long::parseLong)
+                .collect(Collectors.toSet());
+
+        return customerRepository.findAllById(idsSet)
+                .parallelStream()
+                .map(customerMapper::toResponseDto)
+                .filter(customer -> orderId == null || customer.getOrderId().equals(orderId))
+                .toList();
+    }
+
 
     @Transactional(propagation = Propagation.REQUIRED)
     public void addCustomer(String firstName, String lastName) {
@@ -99,24 +191,26 @@ public class CustomerService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public void addOrderToCustomer(Long customerId) {
-        Customer customer = customerRepository.findById(customerId).orElseThrow(
-                () -> new CustomException("The customer doesn't exist"));
+        Optional<Customer> customer1 = customerRepository.findById(customerId);
 
-        kafkaProducer.sendMessage(customerId);
+        if (customer1.isPresent()) {
+            Customer customer = customer1.get();
+            kafkaProducer.sendMessage(customerId);
 
-        orderServiceClient.createOrder(customerId);
+            orderServiceClient.createOrder(customerId);
 
-        customer.setOrderId(customerId);
-        customerRepository.save(customer);
+            customer.setOrderId(customerId);
+            customerRepository.save(customer);
 
-        customerSearchRepository.deleteById(customerId.toString());
-        CustomerDocument customerDocument = new CustomerDocument();
-        customerDocument.setId(customer.getId().toString());
-        customerDocument.setFirstName(customer.getFirstName());
-        customerDocument.setLastName(customer.getLastName());
-        customerDocument.setOrderId(customer.getOrderId());
+            customerSearchRepository.deleteById(customerId.toString());
+            CustomerDocument customerDocument = new CustomerDocument();
+            customerDocument.setId(customer.getId().toString());
+            customerDocument.setFirstName(customer.getFirstName());
+            customerDocument.setLastName(customer.getLastName());
+            customerDocument.setOrderId(customer.getOrderId());
 
-        customerSearchRepository.save(customerDocument);
+            customerSearchRepository.save(customerDocument);
+        }
     }
 
     @Transactional
